fix(template_cache_generator): support traversal of partial files
Fixes #662
